test(ucx-exchange): Cover getUtilization on an uninitialized queue
The maxSize_ == 0 guard added with the OutputBufferManager overrides was
the only new branch without a test. lateTaskCreation already builds the
state it needs -- getData() creates a placeholder queue before
initializeTask(), so its capacity is not known yet -- so assert there
that getUtilization() reports nullopt rather than dividing by zero, and
that isOverutilized() is still false.

Also enqueue a batch before checking utilization in
implementsOutputBufferManager. The previous bounds check passed
trivially on an empty queue and its upper bound was not an invariant the
queue guarantees: producers block only after adding data, so queued
bytes can exceed the maximum.

Stream the output kind through its ostream operator in
UcxOutputQueue::toString() instead of casting it to an int, matching
OutputBuffer::toStringLocked().

+++ velox/experimental/ucx-exchange/UcxQueues.cpp
@@ -555,7 +555,7 @@ void UcxOutputQueue::terminate() {
 std::string UcxOutputQueue::toString() {
   std::lock_guard<std::mutex> l(mutex_);
   std::stringstream out;
-  out << "[UcxOutputQueue: kind=" << static_cast<int>(kind_)
+  out << "[UcxOutputQueue: kind=" << kind_
       << ", numDestinations=" << queues_.size()
       << ", queuedBytes=" << queuedBytes_ << ", maxSize=" << maxSize_
       << ", noMoreQueues=" << noMoreQueues_ << ", atEnd=" << atEnd_
